requirements.txt
SpeechRecognition==3.10.0
gTTS==2.4.0
python-dotenv==1.0.0
google-genai==0.2.2
pygame==2.6.1
streamlit==1.40.0
sounddevice==0.4.6
numpy==1.24.3